pyproject.toml
[tool.poetry]
name = "PoST"
version = "0.0.19-a1"
description = "Reece Innovation | Part of Speech Tagger"
authors = ["Kahlil Wehmeyer <kahlil.wehmeyer@austinai.io>", "Robert Corwin <robert.corwin@austinai.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.7,<3.10.0"
snowflake-connector-python = "3.0.2"
pandas = "^1.5.2"
typer = "^0.7.0"
toml = "^0.10.2"
fastapi = "0.95.0"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
pyspellchecker = "^0.7.0"
contractions = "^0.1.73"
nltk = "^3.7"
fuzzywuzzy = "^0.18.0"
fuzzyset2 = "^0.2.1"
redis = "^4.3.5"
levenshtein = "^0.20.8"
orjson = "^3.8.3"
loguru = "^0.6.0"
elasticsearch-dsl = "^7.4.1"
boto3 = "^1.26.43"
inflect = "^6.0.2"
pyarrow = "8.0.0"
cryptography = "38.0.0"
pyopenssl = "23.0.0"
unicorn = {extras = ["standard"], version = "^2.0.1.post1"}
gunicorn = "^20.1.0"
symspellpy = "^6.7.7"
aws-lambda-powertools = "^2.16.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
isort = "^5.10.1"
black = "^22.10.0"
ipykernel = "^6.17.1"
build = "^0.9.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
coverage-badge = "^1.1.0"
notebook = "^6.5.2"
ruff = "^0.0.254"
ipython = "8.10"



[tool.poetry.group.pos-tagger-image.dependencies]
python = ">=3.8.3,<3.9.7 || >3.9.7,<4.0.0"
symspellpy = "^6.7.7"
orjson = "^3.8.3"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"
pandas = "^1.5.2"
levenshtein = "^0.20.8"
redis = "^4.5.5"
nltk = "^3.8.1"
toml = "^0.10.2"
pydantic = "^1.10.7"
inflect = "^6.0.4"
fuzzywuzzy = "^0.18.0"
fuzzyset2 = "^0.2.2"
contractions = "^0.1.73"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "Q",  "I001", "ERA", "ARG"]
ignore = ["D213", "D203", "E501", "B905", "ARG001"]
line-length = 88 # Same as black
exclude = [
    "post/non-production"
]

[tool.ruff.pydocstyle]
convention = "google"  # Accepts: "google", "numpy", or "pep257".

[tool.ruff.isort]


[tool.ruff.flake8-quotes]
docstring-quotes = "double"
